fix：golangci规范修改part2  (#2995)
* bcs-cluster-reporter规范修改

* bcs-storage规范修改

* bcs-bscp规范修改

* bcs-services/bcs-bscp/cmd/vault-server规范修改

* bcs-user-manager规范修改

* bcs-runtime/bcs-k8s/kubernetes/common规范修改

* /bcs-runtime/bcs-k8s/bcs-component/bcs-unified-apiserver规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-general-pod-autoscaler规范修改

* bcs-netservice-controller、bcs-webhook-server规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-logbeat-sidecar规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-apiserver-prox规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-multi-ns-proxy规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-cluster-autoscaler/bcs-cluster-autoscaler-1.16规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-cluster-autoscaler/bcs-cluster-autoscaler-1.22规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-egress规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-k8s-custom-scheduler规范修改

* bcs-runtime/bcs-k8s/bcs-component/bcs-clusternet-controller规范修改

* bcs-runtime/bcs-k8s/bcs-network规范提交

+++ bcs-services/bcs-bscp/test/benchmark/run/stats.go
@@ -148,7 +148,7 @@ func (s *Statistic) CalculateMetrics() Metrics {
 	var m Metrics
 	m.SustainSeconds = s.SustainSecond
 	m.Concurrent = s.Concurrent
-	m.QPS = float64(float64(len(s.Values)) / s.SustainSecond)
+	m.QPS = float64(len(s.Values)) / s.SustainSecond
 	m.MaxDuration = s.MaxCostDuration
 	m.MinDuration = s.MinCostDuration
